mito-ai-core: move utils to mito-ai-core
