+++ mito-ai-core/mito_ai_core/completions/prompt_builders/agent_execution_prompt.py
@@ -6,20 +6,20 @@ from typing import List
 from mito_ai_core.completions.models import AgentExecutionMetadata
 from mito_ai_core.completions.prompt_builders.prompt_section_registry import SG, Prompt
 from mito_ai_core.completions.prompt_builders.prompt_section_registry.base import PromptSection
+from mito_ai_core.agent.types import AgentExecutionContext
 
-
-def create_agent_execution_prompt(md: AgentExecutionMetadata) -> str:
+def create_agent_execution_prompt(context: AgentExecutionContext, user_input: str) -> str:
     sections: List[PromptSection] = [
         SG.Generic("Reminder", "Remember to choose the correct tool to respond with."),
-        SG.Rules(md.additionalContext),
-        SG.StreamlitAppStatus(md.notebookID, md.notebookPath),
-        SG.Files(md.files),
-        SG.Variables(md.variables),
-        SG.SelectedContext(md.additionalContext),
-        SG.ActiveCellId(md.activeCellId),
-        SG.Notebook(md.aiOptimizedCells),
-        SG.GetCellOutputToolResponse(md.base64EncodedActiveCellOutput),
-        SG.Task(f"{md.input}"),
+        SG.Rules(context.additionalContext),
+        SG.StreamlitAppStatus(context.notebookID, context.notebookPath),
+        SG.Files(context.files),
+        SG.Variables(context.variables),
+        SG.SelectedContext(context.additionalContext),
+        SG.ActiveCellId(context.activeCellId),
+        SG.Notebook(context.aiOptimizedCells),
+        SG.GetCellOutputToolResponse(context.base64EncodedActiveCellOutput),
+        SG.Task(f"{user_input}"),
     ]
 
     prompt = Prompt(sections)
